LayoutTests/media/modern-media-controls/overflow-support/chapters.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->
<meta name="viewport" content="width=device-width">
<script src="../resources/media-controls-utils.js"></script>
<script src="../../../resources/ui-helper.js"></script>
<script src="../../../resources/js-test-pre.js"></script>
<body>
<video src="../../content/long-test.mp4" style="position: absolute; left: 0; top: 0; width: 600px;" controls autoplay muted playsinline>
    <track src="../../content/lorem-ipsum.vtt" kind="chapters">
</video>
<script type="text/javascript">

window.jsTestIsAsync = true;

description("Testing the <code>OverflowSupport</code> behavior for Playback Speed.");

const media = document.querySelector("video");
const shadowRoot = window.internals.shadowRoot(media);

window.internals.setMediaControlsHidePlaybackRates(media, true);

async function checkTimeout(description, promise, ms)
{
    let result = await Promise.race([
        promise.then(() => true),
        UIHelper.delayFor(ms ?? 1000).then(() => false),
    ]);
    if (!result)
        debug(`TIMEOUT: ${description}`);
}

media.addEventListener("play", async function() {
    media.pause();

    await checkTimeout("button added", shouldBecomeDifferent("shadowRoot.querySelector('button.overflow')", "null"));

    await checkTimeout("button visible", shouldBecomeDifferent("shadowRoot.querySelector('button.overflow').getBoundingClientRect().width", "0"));

    await checkTimeout("chapters loaded", shouldBecomeDifferent("media.textTracks[0].mode", "'disabled'"));

    debug("Tapping overflow button...");
    await checkTimeout("clicked button", pressOnElement(shadowRoot.querySelector("button.overflow")));
    await checkTimeout("contextmenu shown", UIHelper.waitForContextMenuToShow());

    debug("Selecting first chapter cue...");
    await checkTimeout("contextmenu action", UIHelper.chooseMenuAction("ipsum"));
    await checkTimeout("contextmenu hidden", UIHelper.waitForContextMenuToHide());

    shouldBeGreaterThanOrEqual("media.currentTime", "5");

    media.remove();
    finishJSTest();
});

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
